test/ui/requirements.txt
allure-pytest
appium-python-client
cryptography
jproperties
pyjwt
pytest
pytz
selenium
webdriver-manager
PyDrive
sshtunnel
pymysql
langdetect